Se copió proyecto de la rama develop a gmatta-2021223

+++ RECIBA/src/user/user.controller.js
@@ -2,11 +2,13 @@
 
 const User = require('./user.model')
 const Range = require('../range/range.model')
+const Recycle = require('../recycler/recycler.model')
 const Achievement = require('../achievements/achievements.model')
 const { encrypt, validateData, check, sensitiveData } = require('../utils/validate')
 const { createToken } = require('../services/jwt')
 const fs = require('fs')
 const path = require('path')
+const { log } = require('console')
 
 const ROLES = Object.freeze({ partner: 'PARTNER', recycler: 'RECYCLER', master: 'MASTER', client: 'CLIENT' })
 
@@ -15,9 +17,10 @@ exports.test = (req, res) => {
 }
 
 /* ----- DEFAULT MASTER ----- */
-exports.defaultMaster = async() => {
+exports.defaultMaster = async () => {
     try {
-        if(!(await User.findOne({ username: 'admin' }))) {
+        let range = await Range.findOne({name: 'ADMIN'})
+        if (!(await User.findOne({ username: 'admin' }))) {
 
             let data = {
                 name: 'admin',
@@ -26,7 +29,8 @@ exports.defaultMaster = async() => {
                 email: 'admin@admin.com',
                 password: 'admin',
                 username: 'admin',
-                role: 'master'
+                role: 'master',
+                range: range._id.toString()
             }
             data.password = await encrypt(data.password)
             let user = new User(data)
@@ -43,7 +47,7 @@ exports.defaultMaster = async() => {
 }
 
 /* ----- LOGIN ----- */
-exports.login = async(req, res) => {
+exports.login = async (req, res) => {
     try {
         let data = req.body
         let params = {
@@ -52,18 +56,21 @@ exports.login = async(req, res) => {
         }
 
         let msg = validateData(params)
-        if(msg) return res.status(400).send({ message: msg })
+        if (msg) return res.status(400).send({ message: msg })
 
         let user = await User.findOne({ username: data.username })
-        if(!user) return res.status(401).send({ message: 'Invalid credentials :(' })
+        if (!user) return res.status(401).send({ message: 'Invalid credentials :(' })
 
-        if(user && await check(data.password, user.password)) {
+        if (user && await check(data.password, user.password)) {
             let token = await createToken(user)
             let logged = {
-                name: user.name,
+                sub: user._id,
                 username: user.username,
                 role: user.role,
-                id: user._id
+                id: user._id,
+                photo: user.photo,
+                exp: user.exp,
+
             }
 
             return res.send({ message: 'Logged!', token: token, user: logged })
@@ -78,10 +85,10 @@ exports.login = async(req, res) => {
 }
 
 /* ----- REGISTER USER ----- */
-exports.register = async(req, res) => {
+exports.register = async (req, res) => {
     try {
         let data = req.body
-        let range = await Range.findOne({name: 'JUNIOR'});
+        let range = await Range.findOne({ name: 'JUNIOR' });
         let params = {
             name: data.name,
             surname: data.surname,
@@ -92,7 +99,7 @@ exports.register = async(req, res) => {
         }
 
         let msg = validateData(params)
-        if(msg) return res.status(400).send({ message: msg })
+        if (msg) return res.status(400).send({ message: msg })
 
         data.role = ROLES.client
         data.password = await encrypt(data.password)
@@ -110,11 +117,11 @@ exports.register = async(req, res) => {
 }
 
 /* ----- GET USERS ----- */
-exports.get = async(req, res) => {
+exports.get = async (req, res) => {
     try {
         let users = await User.find()
 
-        if(!users) return res.status(404).send({ message: 'Users not found :(' })
+        if (!users) return res.status(404).send({ message: 'Users not found :(' })
 
         let data = sensitiveData(users)
         return res.send({ message: 'Users found!', data })
@@ -126,12 +133,47 @@ exports.get = async(req, res) => {
 }
 
 /* ----- GET USER ----- */
-exports.getUser = async(req, res) => {
+exports.getUser = async (req, res) => {
     try {
         let id = req.params.id
 
         let user = await User.findOne({ _id: id })
-        if(!user) return res.status(404).send({ message: 'User not found :(' })
+        if (!user) return res.status(404).send({ message: 'User not found :(' })
+
+        let data = sensitiveData([user])
+
+        return res.send({ message: 'User found!', data })
+
+    } catch (err) {
+        console.error(err)
+        return res.status(500).send({ message: 'Error getting user :(', error: err })
+    }
+}
+
+exports.getOwn = async (req, res) => {
+    try {
+        let id = req.user.sub
+
+        let user = await User.findOne({ _id: id }).populate('range')
+        if (!user) return res.status(404).send({ message: 'User not found :(' })
+
+        let data = sensitiveData([user])
+
+        return res.send({ message: 'User found!', data })
+
+    } catch (err) {
+        console.error(err)
+        return res.status(500).send({ message: 'Error getting user :(', error: err })
+    }
+}
+
+/* ----- GET USERBYUSERNAME ----- */
+exports.getUserByUsername = async (req, res) => {
+    try {
+        let username = req.params.username
+
+        let user = await User.findOne({ username: username }).populate('cart.material')
+        if (!user) return res.status(404).send({ message: 'User not found :(' })
 
         let data = sensitiveData([user])
 
@@ -195,17 +237,18 @@ exports.getUserByUsername = async(req, res) => {
 }
 
 /* -----UPDATE ----- */
-exports.update = async(req, res) => {
+exports.update = async (req, res) => {
     try {
         let data = req.body
 
-        if(data.password) return res.status(400).send({ message: 'Cannot update password here :(' })
-        if(data.role || data.achievements || data.range || data.points) return res.status(400).send({ message: 'Can not update some params' })
-
+        if (data.password) return res.status(400).send({ message: 'Cannot update password here :(' })
+        if (data.role || data.achievements || data.range || data.points ) return res.status(400).send({ message: 'Can not update some params' })
+        let userEmail = await User.findOne({email: data.email})
+        if(userEmail) res.status(418).send({message: 'Can not update the email because is already taked in another User'})
         let upUser = await User.findOneAndUpdate(
-            {_id: req.user.sub},
+            { _id: req.user.sub },
             data,
-            {new: true}
+            { new: true }
         )
 
         return res.send({ message: 'Updated!', user: upUser })
@@ -217,18 +260,23 @@ exports.update = async(req, res) => {
 }
 
 /* -----DELETE ----- */
-exports.del = async(req, res) => {
+exports.del = async (req, res) => {
     try {
         let id = req.user.sub
 
         let user = await User.findOne({ _id: id })
         if (!user) return res.status(404).send({ message: 'User not found to be deleted :(' })
+        if (user.role == ROLES.recycler) {
+            if (
+                await Recycle.findOne({ user: user._id })
+            ) return res.status(401).send({ message: 'Can not delete user because is linked with recycler' });
+        }
         if (user.username === 'admin') return res.status(403).send({ message: 'User MASTER default cannot be deleted' })
         if (user.role === ROLES.admin || user.role === ROLES.master || user.role === ROLES.partner) return res.status(403).send({ message: `User with role "${user.role}" cannot be deleted` })
+        
+        let delUser = await User.findOneAndDelete({ _id: id })
 
-        let delUser = await User.findOneAndDelete({_id: id})
-
-        if(!delUser) return res.status(404).send({ message: 'Account could not be deleted :(' })
+        if (!delUser) return res.status(404).send({ message: 'Account could not be deleted :(' })
         return res.send({ message: 'Account deleted successfully!', error: err })
 
     } catch (err) {
@@ -238,15 +286,16 @@ exports.del = async(req, res) => {
 }
 
 /* -----UPDATE PASSWORD ----- */
-exports.updatePassword = async(req, res) => {
+exports.updatePassword = async (req, res) => {
     try {
+        let id = req.user.sub
         let data = req.body
         let form = {
             password: data.password,
             newPassword: data.newPass
         }
         let msg = validateData(form)
-        if(msg) return res.status(400).send({ message: msg })
+        if (msg) return res.status(400).send({ message: msg })
 
         let user = await User.findOne({ _id: id })
 
@@ -268,10 +317,17 @@ exports.updatePassword = async(req, res) => {
 }
 
 /* -----SAVE ACCOUNT ----- */
-exports.save = async(req, res) => {
-    try {
+exports.save = async (req, res) => {
+     try {
         let data = req.body
         let range = await Range.findOne({name: 'JUNIOR'});
+        data.range = range._id.toString()
+        
+        if(data.role !== ROLES.client){
+            let rangeAdmin = await Range.findOne({name: 'ADMIN'})
+            data.range = rangeAdmin._id.toString()
+        }
+        console.log(data.range);
         let params = {
             name: data.name,
             surname: data.surname,
@@ -280,16 +336,17 @@ exports.save = async(req, res) => {
             password: data.password,
             username: data.username,
             role: data.role,
-            range: range._id
+            range: data.range
         }
 
         let msg = validateData(params)
         if (msg) return res.status(400).send({ message: msg })
-        
         data.password = await encrypt(data.password)
         data.role = data.role.toUpperCase()
-        data.range = range._id.toString()
-
+        if (data.range) {
+            let range = await Range.findOne({ _id: data.range })
+            data.exp = range.initExp
+        }
         let user = new User(data)
         await user.save()
 
@@ -301,20 +358,34 @@ exports.save = async(req, res) => {
     }
 }
 
+
 /* -----UPDATE ACCOUNT ----- */
-exports.updateUser = async(req, res) => {
+exports.updateUser = async (req, res) => {
     try {
         let id = req.params.id
         let data = req.body
         let user = await User.findOne({ _id: id })
-
+        let range = await Range.findOne({ _id: data.range })
+        let rangeAdmin = await Range.findOne({name: 'ADMIN'})
+        let recycler = await Recycle.findOne({ user: user._id}) 
         if (!user) return res.status(404).send({ message: 'User not found!' })
-        if (data.password) return res.status(401).send({ message: 'Cannot update password!' })
-        if (user.role === ROLES.master) return res.status(401).send({ message: 'Cannot update "MASTER"' })
-
-        data.role = data.role.toUpperCase()
-        if (data.role !== ROLES.admin && data.role !== ROLES.client) return res.status(400).send({ message: 'Role not authorized :(' })
-        
+        if(data.role == ROLES.client){
+            if(data.range == null || data.range == undefined || data.range == rangeAdmin._id) 
+            return res.status(400).send({message: `Can not select that range`})
+        }
+        if (user.role == 'RECYCLER' && recycler) {
+            if (
+                (data.role != user.role)
+                ) return res.status(401).send({ message: 'Can not update role because is linked with recycler' });
+            }
+            if (data.password) return res.status(401).send({ message: 'Cannot update password!' })
+            if (user.role === ROLES.master) return res.status(401).send({ message: 'Cannot update "MASTER"' })
+            data.role = data.role.toUpperCase()
+            if(data.role != ROLES.client){
+                data.range = rangeAdmin._id.toString()
+                data.expo = 0
+            }
+        data.exp = range.initExp
         let upUser = await User.findOneAndUpdate(
             { _id: id },
             data,
@@ -322,7 +393,7 @@ exports.updateUser = async(req, res) => {
         )
 
         return res.send({ message: 'Account updated successfully!', user: upUser })
-        
+
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: 'Error while updating account :(', error: err })
@@ -330,16 +401,20 @@ exports.updateUser = async(req, res) => {
 }
 
 /* -----DELETE ACCOUNT ----- */
-exports.delUser = async(req, res) => {
+exports.delUser = async (req, res) => {
     try {
         let id = req.params.id
 
         let user = await User.findOne({ _id: id })
         if (!user) return res.status(404).send({ message: 'User not found' })
+        if (user.role == ROLES.recycler) {
+            if (
+                await Recycle.findOne({ user: user._id })
+            ) return res.status(401).send({ message: 'Can not delete user because is linked with recycler' });
+        }
         if (user.role === ROLES.master) return res.status(401).send({ message: 'Cannot delete user "MASTER"' })
 
         let delUser = await User.findOneAndDelete({ _id: id })
-
         return res.send({ message: 'Account deleted successfully!', user: delUser })
 
     } catch (err) {
@@ -349,13 +424,11 @@ exports.delUser = async(req, res) => {
 }
 
 /* -----UPLOAD PHOTO ----- */
-exports.uploadImg = async(req, res) => {
+exports.uploadImg = async (req, res) => {
     try {
-        const user = req.user
         const id = req.params.id
         const alreadyImg = await User.findOne({ _id: id })
         let pathFile = './src/uploads/users/'
-
         if (alreadyImg.photo) fs.unlinkSync(`${pathFile}${alreadyImg.photo}`)
         if (!req.files.image || !req.files.image.type) return res.status(400).send({ message: 'Have not sent an image :(' })
 
@@ -374,16 +447,17 @@ exports.uploadImg = async(req, res) => {
         ) {
             fs.unlinkSync(filePath)
             return res.status(400).send({ message: 'File extension not admited' })
-        } 
+        }
 
         const upUser = await User.findOneAndUpdate(
             { _id: id },
             { photo: fileName },
             { new: true }
         )
+        
         if (!upUser) return res.status(404).send({ message: 'User not found!' })
         return res.send({ message: 'Photo added successfully' })
-        
+
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: 'Error while uploading img :(', error: err })
@@ -391,7 +465,7 @@ exports.uploadImg = async(req, res) => {
 }
 
 /* -----GET PHOTO ----- */
-exports.getImg = async(req, res) => {
+exports.getImg = async (req, res) => {
     try {
         const fileName = req.params.file
         const pathFile = `./src/uploads/users/${fileName}`
@@ -404,5 +478,34 @@ exports.getImg = async(req, res) => {
     } catch (err) {
         console.error(err)
         return res.status(500).send({ message: 'Error getting img :(', error: err })
+    }
+}
+
+/* ----- CHECK RANGE ----- */
+exports.checkRange = async (req, res) => {
+    try {
+        const id = req.user.sub
+
+        const user = await User.findOne({ _id: id }).populate('range')
+
+        if (user.role !== 'CLIENT') return res.send()
+
+        let limitExp = user.range.limitExp
+
+        if (user.exp >= limitExp) {
+            let range = await Range.findOne({ initExp: limitExp })
+
+            if (!range) return res.send()
+            
+            await User.findOneAndUpdate({ _id: id }, { range: range._id })
+
+            return res.send({ message: `You have been promoted to "${range.name}"`, promoted: true })
+        }
+
+        return res.send({ promoted: false })
+        
+    } catch (err) {
+        console.error(err)
+        return res.status(500).send({ message: 'Error checking range', error: err })
     }
 }
